BI-Projekt-Abgabe/Code/authentic_api_runs/requirements.txt
requests==2.24.0
json5==0.9.5
jsonschema==3.2.0
pandas==1.0.5
pyyaml==5.3.1
geopandas==0.6.3
shapely==1.6.4
fiona==1.8.4
matplotlib==3.2.2


